+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.14.1] - 2026-06-15
-- feat: remove sentry sdk instrumentation from lambdas
+- feat: remove sentry sdk instrumentation from passive agent (tool) lambdas
 
 ## [1.14.0] - 2026-05-27
 - feat: extend `weni run` to support active agents (PreProcessor + Rules)